backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pg8000
python-dotenv
passlib[bcrypt]
python-jose
pydantic
pydantic-settings
psycopg2-binary
bcryptflask
psycopg2-binary
psycopg2-binary
bcrypt
pandas
openpyxl
gunicorn
requests
twilio